Canvas 2.0
About CANVAS

Canvas was designed to be a simple and straightforward WordPress flyout plugin that allows you to display some extra information in a dedicated, full screen area. It’s responsive all the way down to tablets and smartphones, ready for retina screens, and is a great add-on to any theme. It’s super easy to set up and use, but don’t let the clean looks deceive you; the content you’ll display is in no way limited to a predetermined layout.
Freedom to create

Insert text, images, videos, forms etc. as you would into any post or page, then mix, match and format them however you’d like. You are only limited by the tools available in the WordPress editor and any theme add-ons and/or plugins you use on your site. If for example you use shortcodes with your theme, then those can of course be used as well.

As of update 1.4, ‘Canvas’ also fully supports widgets, which can be shown in a single, two, or three columns. As is the rest of ‘Canvas’, the widgetized area is full