Facebook comment slider (Plugin)
Facebook Comment Slider for WordPress offers an opportunity for the visitors to leave comments on your website with this modern, un-intrusive comment box. The plugin also provide a Like Box and Fan Page Wall features to extend your website with this All-in-One Facebook Plugin.

You don’t need to integrate the Like and Share button to the posts or pages, because it is already contain and it can be appear on every pages and posts. The visitors will find the Like button, Share button and the Comments Box on the Facebook Slider. 
You can enable, disable and set time of the shake effect.

INCLUDED FACEBOOK LIKEBOX, COMMENT BOX AND WALL POSTS

The plugin contains three different slider to help you to integrate all Facebook marketing opportunities to your website. You can display the Like Box on the homepage only or display all of three sliders on every pages. Simple shortcodes gives you the full control to achieve your goals.


codecanyon-5166547-facebook-comment-slider